Refactor

+++ portfolio-client/src/components/adminComponents/About.tsx
@@ -1,27 +1,23 @@
-import { useState, useEffect } from 'react';
+import { useState, useContext } from 'react';
 import './AboutAdmin.css';
+import { AppContext } from '../../App';
 
-type Props = {
-  accessAdmin: boolean;
+interface IAboutProps {
+  token: string;
 }
 
-const About = (props: Props) => {
+const About = (props: IAboutProps) => {
 
-  const [aboutParagraphs, setAboutParagraphs] = useState([{ id: 0, text: '' }]);
-  const [aboutParagraphTexts, setAboutParagraphsTexts] = useState(['']);
-  const [homeLinks, setHomeLinks] = useState([{ imgUrl: '', url: '', text: '' }]);
+  const adminAccess = props.token !== 'Not Authorized';
 
-  useEffect(() => {
-    fetch('https://jeportapi.azurewebsites.net/api/about/')
-      .then(response => response.json())
-      .then(result => { setAboutParagraphs(result); setAboutParagraphsTexts(result.map((p: { id: number, text: string }) => p.text)) })
-  }, [])
+  const aboutProps = useContext(AppContext).aboutProps;
+  const aboutParagraphs = aboutProps.aboutParagraphs;
+  const setAboutParagraphs = aboutProps.setAboutParagraphs;
+  const homeLinks = aboutProps.homeLinks;
+  const setHomeLinks = aboutProps.setHomeLinks;
 
-  useEffect(() => {
-    fetch('https://jeportapi.azurewebsites.net/api/home/home-links')
-      .then(response => response.json())
-      .then(result => setHomeLinks(result))
-  }, [])
+  const [aboutParagraphTexts, setAboutParagraphsTexts] = useState(aboutParagraphs.map(p => p.text));
+  const [homeLinkTexts, setHomeLinkTexts] = useState(homeLinks.map(l => l.text))
 
   const deleteParagraph = (index: number) => {
     setAboutParagraphsTexts(
@@ -31,66 +27,85 @@ const About = (props: Props) => {
     )
   }
 
-  const postAboutParagraph = ( text: string ) => {
-    fetch('https://jeportapi.azurewebsites.net/api/about', {
+  const postAboutParagraph = (text: string) => {
+    fetch('https://jeportapi.azurewebsites.net/api/About', {
       method: 'POST',
       headers: {
-        'Content-Type': 'application/json'
+        'Content-Type': 'application/json',
+        'Authorization': `Bearer ${props.token}`
       },
       body: JSON.stringify({ text })
     })
   }
 
   const putAboutParagraph = (id: number, text: string) => {
-    fetch('https://jeportapi.azurewebsites.net/api/about', {
+    fetch('https://jeportapi.azurewebsites.net/api/About', {
       method: 'PUT',
       headers: {
-        'Content-Type': 'application/json'
+        'Content-Type': 'application/json',
+        'Authorization': `Bearer ${props.token}`
       },
       body: JSON.stringify({ id, text })
     })
   }
 
   const deleteAboutParagraph = (id: number) => {
-    fetch(`https://jeportapi.azurewebsites.net/api/about/${id}`, {
-      method: 'DELETE'
+    fetch(`https://jeportapi.azurewebsites.net/api/About/${id}`, {
+      method: 'DELETE',
+      headers: {
+        'Authorization': `Bearer ${props.token}`
+      }
     })
   }
 
   const handleChanges = () => {
-    
+
     const currentParagraphsLength = aboutParagraphs.length;
     const updatedParagraphsLength = aboutParagraphTexts.length;
     const difference = updatedParagraphsLength - currentParagraphsLength;
 
-    if (difference < 0) {
-      for (let i = 0; i < updatedParagraphsLength; i++) {
-        if (aboutParagraphs[i].text !== aboutParagraphTexts[i]) {
-          putAboutParagraph(aboutParagraphs[i].id, aboutParagraphTexts[i]);
+    interface IAboutParagraph {
+      id: number,
+      text: string
+    };
+
+    if (adminAccess) {
+      if (difference < 0) {
+        for (let i = 0; i < updatedParagraphsLength; i++) {
+          if (aboutParagraphs[i].text !== aboutParagraphTexts[i]) {
+            putAboutParagraph(aboutParagraphs[i].id, aboutParagraphTexts[i]);
+          }
+        }
+        for (let i = updatedParagraphsLength; i < currentParagraphsLength; i++) {
+          deleteAboutParagraph(aboutParagraphs[i].id);
         }
       }
-      for (let i = updatedParagraphsLength; i < currentParagraphsLength; i++) {
-        deleteAboutParagraph(aboutParagraphs[i].id);
+
+      if (difference === 0) {
+        for (let i = 0; i < updatedParagraphsLength; i++) {
+          if (aboutParagraphs[i].text !== aboutParagraphTexts[i]) {
+            putAboutParagraph(aboutParagraphs[i].id, aboutParagraphTexts[i]);
+          }
+        }
+      }
+
+      if (difference > 0) {
+        for (let i = 0; i < currentParagraphsLength; i++) {
+          if (aboutParagraphs[i].text !== aboutParagraphTexts[i]) {
+            putAboutParagraph(aboutParagraphs[i].id, aboutParagraphTexts[i]);
+          }
+        }
+        for (let i = currentParagraphsLength; i < updatedParagraphsLength; i++) {
+          postAboutParagraph(aboutParagraphTexts[i]);
+        }
       }
     }
-
-    if (difference === 0) {
-      for (let i = 0; i < updatedParagraphsLength; i++) {
-        if (aboutParagraphs[i].text !== aboutParagraphTexts[i]) {
-          putAboutParagraph(aboutParagraphs[i].id, aboutParagraphTexts[i]);
-        }
-      }
-    }
-
-    if (difference > 0) {
-      for (let i = 0; i < currentParagraphsLength; i++) {
-        if (aboutParagraphs[i].text !== aboutParagraphTexts[i]) {
-          putAboutParagraph(aboutParagraphs[i].id, aboutParagraphTexts[i]);
-        }
-      }
-      for (let i = currentParagraphsLength; i < updatedParagraphsLength; i++) {
-        postAboutParagraph(aboutParagraphTexts[i]);
-      }
+    else {
+      const newAboutParagraphs: IAboutParagraph[] = [];
+      aboutParagraphTexts.forEach((value, index) =>
+        newAboutParagraphs.push({ id: index, text: value })
+      )
+      setAboutParagraphs(newAboutParagraphs);
     }
   }
 
@@ -98,26 +113,49 @@ const About = (props: Props) => {
     <div className="About-content">
       <form>
         {aboutParagraphTexts.map((aboutParagraph, paragraphIndex) =>
-        <div>
-          <textarea
-            value={aboutParagraph}
-            onChange={e => setAboutParagraphsTexts(aboutParagraphTexts.map((p, textIndex) => {
-              if (paragraphIndex === textIndex) {
-                p = e.target.value;
-              }
-              return p;
-            }))}
-            rows={5}
-            cols={100}
-          />
-          <button type="button" onClick={() => deleteParagraph(paragraphIndex)}>Delete</button>
-        </div>
+          <div key={paragraphIndex}>
+            <textarea
+              value={aboutParagraph}
+              onChange={e => setAboutParagraphsTexts(aboutParagraphTexts.map((p, textIndex) => {
+                if (paragraphIndex === textIndex) {
+                  p = e.target.value;
+                }
+                return p;
+              }))}
+              rows={5}
+              cols={100}
+            />
+            <button type="button" onClick={() => deleteParagraph(paragraphIndex)}>Delete</button>
+          </div>
         )}
-        <button type="button" onClick={() => setAboutParagraphsTexts([...aboutParagraphTexts, '']) }>Add Paragraph</button>
-        <button type="button" onClick={() => { if (props.accessAdmin) { handleChanges() } } }>Save Changes</button>
+        <button type="button" onClick={() => setAboutParagraphsTexts([...aboutParagraphTexts, ''])}>Add Paragraph</button>
+        <button type="button" onClick={() => { if (props.token) { handleChanges() } }}>Save Changes</button>
+      </form>
+      <form>
+        {homeLinkTexts.map((homeLinkText, index) => 
+          <div key={index}>
+            <img src={homeLinks[index].imgUrl} alt="logo" height="50rem" width="auto" />
+            <textarea
+              value={homeLinkText}
+              onChange={e => setHomeLinkTexts(homeLinkTexts.map((l, textindex) => {
+                if (index === textindex) {
+                  l = e.target.value
+                }
+                return l;
+              }))}
+              rows={2}
+              cols={100}
+            />
+          </div>
+        )}
       </form>
     </div>
   )
 }
 
 export default About;
+
+
+//{
+
+//}
